Refactor blobstore client creation
- collapse BaseClient and Client
- Blobstores is responsible for creating specific client based on
  provider string
