+++ src/bosh-director/lib/bosh/director/blobstore/retryable_blobstore_client.rb
@@ -2,7 +2,7 @@ require 'forwardable'
 
 module Bosh::Director
   module Blobstore
-    class RetryableBlobstoreClient < BaseClient
+    class RetryableBlobstoreClient < Client
       extend Forwardable
 
       def initialize(client, retryable)
