+++ src/bosh-director/spec/unit/blobstore/retryable_blobstore_client_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 module Bosh::Director::Blobstore
   describe RetryableBlobstoreClient do
     subject { described_class.new(wrapped_client, retryable) }
-    let(:wrapped_client) { instance_double('Bosh::Director::Blobstore::BaseClient') }
+    let(:wrapped_client) { instance_double('Bosh::Director::Blobstore::Client') }
     let(:retryable)      { Bosh::Retryable.new(tries: 2, sleep: 0, on: [BlobstoreError]) }
 
     it_calls_wrapped_client_methods(except: [:get])
