+++ src/bosh-director/spec/unit/blobstore/sha1_verifiable_blobstore_client_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 module Bosh::Director::Blobstore
   describe Sha1VerifiableBlobstoreClient do
     subject { described_class.new(wrapped_client, per_spec_logger) }
-    let(:wrapped_client) { instance_double('Bosh::Director::Blobstore::BaseClient') }
+    let(:wrapped_client) { instance_double('Bosh::Director::Blobstore::Client') }
     let(:multidigest_path) { 'some/path/to/binary' }
 
     it_calls_wrapped_client_methods(except: [:get])
