+++ src/bosh-director/spec/unit/deployment_plan/stages/package_compile_stage_spec.rb
@@ -128,7 +128,7 @@ module Bosh::Director
 
     let(:compiled_package_finder) { DeploymentPlan::CompiledPackageFinder.new(per_spec_logger) }
 
-    let(:blobstore) { instance_double(Bosh::Director::Blobstore::BaseClient) }
+    let(:blobstore) { instance_double(Bosh::Director::Blobstore::Client) }
 
     let(:director_test_certs) { "these\nare\nthe\ncerts" }
     let(:director_test_certs_sha1) { ::Digest::SHA1.hexdigest director_test_certs }
