+++ src/bosh-director/spec/unit/logs_fetcher_spec.rb
@@ -18,7 +18,7 @@ module Bosh::Director
       filters
     end
     let(:log) { double('logger') }
-    let(:blobstore) { instance_double(Bosh::Director::Blobstore::BaseClient) }
+    let(:blobstore) { instance_double(Bosh::Director::Blobstore::Client) }
 
     before do
       allow(App).to receive_message_chain(:instance, :blobstores, :blobstore).and_return(blobstore)
